new boot

+++ object/boot/src/AVLTree.o.c
@@ -660,10 +660,10 @@ static oop AVLTreeNode__delta(oop v__closure, oop v_stateful_self, oop v_self)
   _line(77);
   _2=_sendv(s_avlTreeNodeHeight, 1, _2);
   {
-    int _l= (int)_1 >> 1;
-    int _r= (int)_2 >> 1;
+    int _l= (long)_1 >> 1;
+    int _r= (long)_2 >> 1;
     int _s= (_l - _r);
-    if ((1 & (int)_1 & (int)_2) && ((_s ^ (_s << 1)) >= 0))  _1= (oop)(_s << 1 | 1);  else _1= _sendv(s__2d, 2, _1, _2);
+    if ((1 & (long)_1 & (long)_2) && ((_s ^ (_s << 1)) >= 0))  _1= (oop)(long)(_s << 1 | 1);  else _1= _sendv(s__2d, 2, _1, _2);
   }
   _leave();
   return _1;
@@ -687,8 +687,8 @@ static oop AVLTreeNode__balance(oop v__closure, oop v_stateful_self, oop v_self)
   _1= 0;
   _2= v_delta;
   _3= l_19;
-  if (1 & (int)_2) {
-    _2= (((int)_2 < (int)_3) ? v_true : v_false);
+  if (1 & (long)_2) {
+    _2= (((long)_2 < (long)_3) ? v_true : v_false);
   } else _2= _sendv(s__3c, 2, _2, _3);
   if (!_2) goto _l6;
  {
@@ -698,8 +698,8 @@ static oop AVLTreeNode__balance(oop v__closure, oop v_stateful_self, oop v_self)
   _line(85);
   _2=_sendv(s_delta, 1, _2);
   _3= l_20;
-  if (1 & (int)_2) {
-    _2= (((int)_2 > (int)_3) ? v_true : v_false);
+  if (1 & (long)_2) {
+    _2= (((long)_2 > (long)_3) ? v_true : v_false);
   } else _2= _sendv(s__3e, 2, _2, _3);
   if (!_2) goto _l7;
  {
@@ -720,8 +720,8 @@ static oop AVLTreeNode__balance(oop v__closure, oop v_stateful_self, oop v_self)
   _1= 0;
   _2= v_delta;
   _3= l_21;
-  if (1 & (int)_2) {
-    _2= (((int)_2 > (int)_3) ? v_true : v_false);
+  if (1 & (long)_2) {
+    _2= (((long)_2 > (long)_3) ? v_true : v_false);
   } else _2= _sendv(s__3e, 2, _2, _3);
   if (!_2) goto _l8;
  {
@@ -731,8 +731,8 @@ static oop AVLTreeNode__balance(oop v__closure, oop v_stateful_self, oop v_self)
   _line(89);
   _2=_sendv(s_delta, 1, _2);
   _3= l_20;
-  if (1 & (int)_2) {
-    _2= (((int)_2 < (int)_3) ? v_true : v_false);
+  if (1 & (long)_2) {
+    _2= (((long)_2 < (long)_3) ? v_true : v_false);
   } else _2= _sendv(s__3c, 2, _2, _3);
   if (!_2) goto _l9;
  {
@@ -763,8 +763,8 @@ static oop AVLTreeNode__balance(oop v__closure, oop v_stateful_self, oop v_self)
   _line(92);
   _2=_sendv(s_height, 1, _2);
   _3= ((struct t_AVLTreeNode *)v_stateful_self)->v_height;
-  if (1 & (int)_2 & (int)_3) {
-    _2= (((int)_2 > (int)_3) ? v_true : v_false);
+  if (1 & (long)_2 & (long)_3) {
+    _2= (((long)_2 > (long)_3) ? v_true : v_false);
   } else _2= _sendv(s__3e, 2, _2, _3);
  }
  _l11:;
@@ -788,8 +788,8 @@ static oop AVLTreeNode__balance(oop v__closure, oop v_stateful_self, oop v_self)
   _line(93);
   _2=_sendv(s_height, 1, _2);
   _3= ((struct t_AVLTreeNode *)v_stateful_self)->v_height;
-  if (1 & (int)_2 & (int)_3) {
-    _2= (((int)_2 > (int)_3) ? v_true : v_false);
+  if (1 & (long)_2 & (long)_3) {
+    _2= (((long)_2 > (long)_3) ? v_true : v_false);
   } else _2= _sendv(s__3e, 2, _2, _3);
  }
  _l13:;
@@ -804,10 +804,10 @@ static oop AVLTreeNode__balance(oop v__closure, oop v_stateful_self, oop v_self)
   _1= ((struct t_AVLTreeNode *)v_stateful_self)->v_height;
   _2= l_21;
   {
-    int _l= (int)_1 >> 1;
-    int _r= (int)_2 >> 1;
+    int _l= (long)_1 >> 1;
+    int _r= (long)_2 >> 1;
     int _s= _l + _r;
-    if ((1 & (int)_1) && ((_s ^ (_s << 1)) >= 0))  _1= (oop)(_s << 1 | 1);  else _1= _sendv(s__2b, 2, _1, _2);
+    if ((1 & (long)_1) && ((_s ^ (_s << 1)) >= 0))  _1= (oop)(long)(_s << 1 | 1);  else _1= _sendv(s__2b, 2, _1, _2);
   }
   ((struct t_AVLTreeNode *)v_stateful_self)->v_height= _1;
   _1= v_self;
@@ -1193,8 +1193,8 @@ static oop b_41(oop v__closure, oop v__self, oop v_a, oop v_b)
   _line(209);
   _1= v_a;
   _2= v_b;
-  if (1 & (int)_1 & (int)_2) {
-    _1= (((int)_1 < (int)_2) ? v_true : v_false);
+  if (1 & (long)_1 & (long)_2) {
+    _1= (((long)_1 < (long)_2) ? v_true : v_false);
   } else _1= _sendv(s__3c, 2, _1, _2);
   _leave();
   return _1;
