+++ object/boot/src/Association.o.c
@@ -340,8 +340,8 @@ static oop Association___3d(oop v__closure, oop v_stateful_self, oop v_self, oop
   _2= v_anAssociation;
   _line(33);
   _2=_sendv(s_value, 1, _2);
-  if (1 & (int)_1 & (int)_2) {
-    _1= (((int)_1 == (int)_2) ? v_true : v_false);
+  if (1 & (long)_1 & (long)_2) {
+    _1= (((long)_1 == (long)_2) ? v_true : v_false);
   } else _1= _sendv(s__3d, 2, _1, _2);
  }
  _l1:;
@@ -362,8 +362,8 @@ static oop Association__hash(oop v__closure, oop v_stateful_self, oop v_self)
   _2= ((struct t_Association *)v_stateful_self)->v_value;
   _line(34);
   _2=_sendv(s_hash, 1, _2);
-  if ((1 & (int)_1 & (int)_2) && (_1 > 0) && (_2 > 0)) {
-    _1= (oop)(((int)_1 ^ (int)_2) | 1);
+  if ((1 & (long)_1 & (long)_2) && (_1 > 0) && (_2 > 0)) {
+    _1= (oop)(((long)_1 ^ (long)_2) | 1);
   } else _1= _sendv(s_bitXor_, 2, _1, _2);
   _leave();
   return _1;
