+++ object/boot/src/CManagedCodeGenerator.o.c
@@ -2222,8 +2222,8 @@ static oop CManagedCodeGenerator__genType_sequence_method_(oop v__closure, oop v
   _1= 0;
   _2= v_sequenceNumber;
   _3= l_83;
-  if (1 & (int)_2) {
-    _2= (((int)_2 > (int)_3) ? v_true : v_false);
+  if (1 & (long)_2) {
+    _2= (((long)_2 > (long)_3) ? v_true : v_false);
   } else _2= _sendv(s__3e, 2, _2, _3);
   if (!_2) goto _l9;
  {
@@ -2598,16 +2598,16 @@ static oop b_140(oop v__closure, oop v__self)
   /* and: */
   _1= v_ascii;
   _2= l_142;
-  if (1 & (int)_1) {
-    _1= (((int)_1 >= (int)_2) ? v_true : v_false);
+  if (1 & (long)_1) {
+    _1= (((long)_1 >= (long)_2) ? v_true : v_false);
   } else _1= _sendv(s__3e_3d, 2, _1, _2);
   if (!_1) goto _l14;
  {
   /* and: */
   _1= v_ascii;
   _2= l_143;
-  if (1 & (int)_1) {
-    _1= (((int)_1 < (int)_2) ? v_true : v_false);
+  if (1 & (long)_1) {
+    _1= (((long)_1 < (long)_2) ? v_true : v_false);
   } else _1= _sendv(s__3c, 2, _1, _2);
   if (!_1) goto _l15;
  {
@@ -2683,8 +2683,8 @@ static oop b_136(oop v__closure, oop v__self, oop v_char)
   _line(372);
   _1= ((oop *)_state2)[1];  /* char */
   _2= l_138;
-  if (1 & (int)_1 & (int)_2) {
-    _1= (((int)_1 == (int)_2) ? v_true : v_false);
+  if (1 & (long)_1 & (long)_2) {
+    _1= (((long)_1 == (long)_2) ? v_true : v_false);
   } else _1= _sendv(s__3d, 2, _1, _2);
   /* Scope() */
   /* Scope('char'->ArgumentVariableNode) */
@@ -3749,8 +3749,8 @@ static oop CManagedCodeGenerator__genFreeState_outer_offset_(oop v__closure, oop
   /* ifTrue:ifFalse: */
   _1= v_indir;
   _2= l_275;
-  if (1 & (int)_1) {
-    _1= (((int)_1 == (int)_2) ? v_true : v_false);
+  if (1 & (long)_1) {
+    _1= (((long)_1 == (long)_2) ? v_true : v_false);
   } else _1= _sendv(s__3d, 2, _1, _2);
   if (!_1) goto _l29;
  {
@@ -3770,10 +3770,10 @@ static oop CManagedCodeGenerator__genFreeState_outer_offset_(oop v__closure, oop
   _1= v_indir;
   _2= l_277;
   {
-    int _l= (int)_1 >> 1;
-    int _r= (int)_2 >> 1;
+    int _l= (long)_1 >> 1;
+    int _r= (long)_2 >> 1;
     int _s= (_l - _r);
-    if ((1 & (int)_1) && ((_s ^ (_s << 1)) >= 0))  _1= (oop)(_s << 1 | 1);  else _1= _sendv(s__2d, 2, _1, _2);
+    if ((1 & (long)_1) && ((_s ^ (_s << 1)) >= 0))  _1= (oop)(long)(_s << 1 | 1);  else _1= _sendv(s__2d, 2, _1, _2);
   }
   /* Scope() */
   /* Scope('indir'->ArgumentVariableNode 'compiler'->SlotVariableNode 'encoder'->SlotVariableNode 'stateful_self'->ArgumentVariableNode 'output'->SlotVariableNode 'scopeTag'->ArgumentVariableNode 'self'->ArgumentVariableNode 'offset'->ArgumentVariableNode 'nextLabel'->SlotVariableNode) */
@@ -3790,10 +3790,10 @@ static oop CManagedCodeGenerator__genFreeState_outer_offset_(oop v__closure, oop
   _1= v_indir;
   _2= l_277;
   {
-    int _l= (int)_1 >> 1;
-    int _r= (int)_2 >> 1;
+    int _l= (long)_1 >> 1;
+    int _r= (long)_2 >> 1;
     int _s= (_l - _r);
-    if ((1 & (int)_1) && ((_s ^ (_s << 1)) >= 0))  _1= (oop)(_s << 1 | 1);  else _1= _sendv(s__2d, 2, _1, _2);
+    if ((1 & (long)_1) && ((_s ^ (_s << 1)) >= 0))  _1= (oop)(long)(_s << 1 | 1);  else _1= _sendv(s__2d, 2, _1, _2);
   }
   /* Scope() */
   /* Scope('indir'->ArgumentVariableNode 'compiler'->SlotVariableNode 'encoder'->SlotVariableNode 'stateful_self'->ArgumentVariableNode 'output'->SlotVariableNode 'scopeTag'->ArgumentVariableNode 'self'->ArgumentVariableNode 'offset'->ArgumentVariableNode 'nextLabel'->SlotVariableNode) */
@@ -4552,10 +4552,10 @@ static oop CManagedCodeGenerator__createStateVector_inScope_constructor_(oop v__
   _2= v_size;
   _3= l_387;
   {
-    int _l= (int)_2 >> 1;
-    int _r= (int)_3 >> 1;
+    int _l= (long)_2 >> 1;
+    int _r= (long)_3 >> 1;
     int _s= _l + _r;
-    if ((1 & (int)_2) && ((_s ^ (_s << 1)) >= 0))  _2= (oop)(_s << 1 | 1);  else _2= _sendv(s__2b, 2, _2, _3);
+    if ((1 & (long)_2) && ((_s ^ (_s << 1)) >= 0))  _2= (oop)(long)(_s << 1 | 1);  else _2= _sendv(s__2b, 2, _2, _3);
   }
   _sendv(s_print_, 2, _1, _2);
   _2= l_388;
@@ -4889,10 +4889,10 @@ static oop CManagedCodeGenerator__pushLabel(oop v__closure, oop v_stateful_self,
   _1= ((struct t_CManagedCodeGenerator *)v_stateful_self)->v_nextLabel;
   _2= l_435;
   {
-    int _l= (int)_1 >> 1;
-    int _r= (int)_2 >> 1;
+    int _l= (long)_1 >> 1;
+    int _r= (long)_2 >> 1;
     int _s= _l + _r;
-    if ((1 & (int)_1) && ((_s ^ (_s << 1)) >= 0))  _1= (oop)(_s << 1 | 1);  else _1= _sendv(s__2b, 2, _1, _2);
+    if ((1 & (long)_1) && ((_s ^ (_s << 1)) >= 0))  _1= (oop)(long)(_s << 1 | 1);  else _1= _sendv(s__2b, 2, _1, _2);
   }
   ((struct t_CManagedCodeGenerator *)v_stateful_self)->v_nextLabel= _1;
   _leave();
